Type Validation
Was curious about implementing a JS implementation for going back and forth from Pig latin and English, and found this implementation! Going to learn how he made it work here, then maybe make my own ESM implementation for it as well. Or maybe I could contribute/upstream it back to this project as well :)

https://www.google.com/search?q=pig+latin