Demo Concept
Trying to test whether it's fully reversible in it's current state.

I'm also wondering whether you can detect if a string is a valid non-Pig Latin/valid Pig Latin one, then you can losslessly go between the two, with the magic of being able to determine which one it is. So it would be really cool to be able to either convert one way or the other, as well as detect whether it's valid Pig Latin or not. So maybe it's a igpay interlay ahahay.