Uppercase Distinction Handling
I want to properly account for first-letter uppercased words, by making the new word cased the same, and removing the casing on the original first letter once it is moved to the other half of the derived one.

I also want to add support for handling special characters like punctuation, since currently they are included in the word groups themselves, which breaks things a bit.

I tried making my `isUpperCase()` function generic, but it didn't work unfortunately. I wanted it to return a `letter is Uppercase<T>` kind of thing, where `T` is the `letter` value parameter, and it lexically scopes the uppercase state to where you are checking that value. That stuff is super cool to me, I really like that kind of concept.

Lucy Has the Grip of a Crop Duster - MFTJ