Naming Similarities
The converting to/from functions are actually very similar in how they work, I think they could possibly be derived from a single implementation if need be, but I'm not totally sure if that's the best idea. But for this commit at least, parts of both of the functions, in the middle, are symmetrical in their logic and naming now.

Just heard about Field Music from a Mike Keneally interview, sounds like it's a band that Marco Minnemann really liked(s) a bunch. Should I get it for Bandcamp Friday tomorrow? :D Might help with my music drought I've slipped into again recently. I think it takes about a month or two for me to be in search of new music again, very interesting.
https://fieldmusic.bandcamp.com/album/commontime-2