All-Letter Case Mapping
Nearly working! Want to allow any-cased word to be transformed both directions, and it still look formatted nicely, rather than only working with the first letter if it is uppercase.

It still needs a bit of work for handling punctuation, as well as the hyphens for each group. I want to properly merge them into a single word, rather than have hyphens in them. Since you can parse the word backwards when converting back to regular English, you can calculate where the original group was when going the other direction as well, without needing a separator. If it were a more complex pattern, it may not be possible, but it is here. Hmm, wait, I'm not completely sure about that actually. What if the word starts and ends with two consonants? Maybe this won't be possible, we'll see.

Looked into a heck ton of RegExp patterns to help work on this setup, used a few of them to figure out this commit. Most of the others are for the parts I mentioned above. I didn't have internet at the Cafe today, so I didn't get to look into things like that until now.

This last month was so great for running! Did a bunch of hills, and did not too much distance. I think I will relax a bit more this month, maybe work on some more coding and music, as I've been wanting to get more time with those again lately. You know when you have things you want to do, but can't quite yet? aaagh! Nice to be on the tail end of that hopefully.

I am really liking Field Music, I think I will get it tomorrow :)

https://stackoverflow.com/questions/8462619/regular-expression-first-character-matching-a-z
https://stackoverflow.com/questions/5516119/regular-expression-to-match-characters-at-beginning-of-line-only
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String/match
https://superuser.com/questions/749164/how-to-use-regex-to-capitalise-the-first-letter-of-each-word-in-a-sentence
https://stackoverflow.com/questions/8274562/regex-to-check-if-the-first-character-is-uppercase (Wanted to see if I could improve my `isUpperCase()` function)
https://stackoverflow.com/questions/9862761/how-to-check-if-character-is-a-letter-in-javascript (Part of the one above, I wanted to ensure that only alphabetical characters count; wait! realization! I can just invert it's use to `toUpperCase()` calls instead, as that's what non-alphabetic characters are counted as! gonna have to try that, now that makes more sense again why most checks use `toUpperCase()` instead of `toLowerCase()`; wait I might be wrong, I'll be back to check that. Forgot how my implementation currently uses that, lol)
https://stackoverflow.com/questions/13353663/what-is-the-regular-expression-to-allow-uppercase-lowercase-alphabetical-charac
https://stackoverflow.com/questions/5871793/remove-punctuation-from-string-with-regex (The current issue is that punctuation-based characters are unintentionally mapping their indexes incorrectly in the write phase from the looks of it, so I have to figure out how to properly map things to account for the current use of the hyphen separator)
https://stackoverflow.com/questions/2788112/what-is-the-regular-expression-for-not-whitespace-and-not-a-hyphen

I want to move the `vowels` array over to a RegExp instead as well.