Vowels RegExp + Diffed Formatting
Additional formatting that I didn't add yet to help with the diff a bit better.

Wasn't sure the difference between `\w` and `[a-z]`, looks like `\w` encompasses `[a-z]`, and is alphanumeric too, so maybe even more what I want.

https://www.google.com/search?q=%5Cw+vs+a-z

Forgot to include this demo in the previous commit message. I was trying to figure out why the screaming demo wasn't working, I thought this could be a fix for the `*Group` variables (more specifically the `upperCaseMap` and `result`), but it didn't help fix it, as the group ones don't have the hyphen used anywhere, but something about it in the process is still causing casing issues.

```js
"Hi-Hey-Ha-Ho!".replace("-","")
```