Zero-Width Space Separator
Not sure how I feel about this, but it's an idea! It properly conserves where the break is in the text, using a hidden character instead. It's a bit odd when selecting it though, so I'm not completely sold on it. It does allow properly going back and forth from visibly-seamless sentences, but you have to ensure the sentence has proper Unicode breaks in the sections as well, so I'm not sure about it completely.

https://stackoverflow.com/questions/17978720/invisible-characters-ascii
https://en.wikipedia.org/wiki/Zero-width_space
https://dev.to/virenb/solving-pig-latin-freecodecamp-algorithm-challenges-2bk0
https://stackoverflow.com/questions/69756937/javascript-pig-latin-converter-optimizations
https://stackoverflow.com/questions/7055600/u200b-zero-width-space-characters-in-my-js-code-where-did-they-come-from
https://stackoverflow.com/questions/19936374/is-there-an-invisible-character-that-is-not-regarded-as-whitespace