`NULL` Delimiter
I think this may be a possibly better solution. But I'm not sure? This seems to be a better-supported character, or rather it gets rendered properly in things like the terminal, which `\200b` wasn't (it was being shown as a space still).

I think some kind of delimiter *has* to be there, if one has the intention of programatically reversing the Pig Latin, without a database or AI of some sort to figure out the makeup of the word, and what it's inverse is. I want to be able to go back and forth seamlessly without needing to have a collection of words to figure out what it may be. It would be nice if it didn't need a delimiter too though, as then you could fully have a cooler way to store secret messages without having as much information about how it was previously arranged as.

I want to make a custom NBT format that uses Pig Latin to serialize it's keys and values, along with VarInt and a compression algorithm of some sort hahaha. Maybe that could be part of my demo for the JSON-replacer implementation, using Pig Latin in the replacer demo/test haha.

https://superuser.com/questions/1746272/how-to-remove-all-zero-width-space-from-the-text
https://stackoverflow.com/questions/18306574/special-unicode-escape-characters-to-hide-text
https://en.wikipedia.org/wiki/Unicode_control_characters
https://superuser.com/questions/1565894/how-to-remove-what-are-apparently-control-characters-in-notepad
https://en.wikipedia.org/wiki/Null_character
https://superuser.com/questions/1537462/what-is-the-purpose-of-these-nul-characters
https://stackoverflow.com/questions/6907297/can-utf-8-contain-zero-byte

I looked into those last few articles because I wanted to see if there is any unintentional overlap if I am to use a control character of some sort, as I don't want it to cause side-effects of meaning something else that I don't mean for it to.