Vowel-Based Suffix Handling
It wasn't currently setup to handle this yet, so now it uses the other suffixes that it might need, like for words that use a vowel for the first letter. It's not symmetrical yet though, and things are still broken with the "yelling" demo, which I was also trying to figure out and debug with this one.

https://www.getblend.com/blog/how-to-speak-pig-latin/
https://en.wikipedia.org/wiki/Pig_Latin